% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Country Population.}
\usage{
CountryPopulation(countryName)
}
\arguments{
\item{countryName}{\itemize{
\item a name of a country
}}
}
\value{
ggplot graph of that country's population data
}
\description{
This function takes in a string of a country name and returns a graph of its
population data from the dataset World_Population from this package.
}
\examples{
CountryPopulation( "China" )
CountryPopulation( "Italy" )
}
