% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup_Data.R
\docType{data}
\name{World_Cup_Data}
\alias{World_Cup_Data}
\title{World Cup data}
\format{
A data frame with 22 observations with 4 columns.
\describe{
\item{WorldCup}{The country and year of each World Cup;                 Unit: none;     Datatype: chr}
\item{Matches}{The number of matches in the specific World Cup;         Unit: match;    Datatype: num}
\item{Totalattendance}{Total attendance of the specific World Cup;      Unit: person;   Datatype: num}
\item{Averageattendance}{Average attendance of the specific World Cup;  Unit: person;   Datatype: num}
}
}
\usage{
World_Cup_Data
}
\description{
A set of data about the past World Cups attendance.
}
\keyword{datasets}
