% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Population.R
\docType{data}
\name{World_Population}
\alias{World_Population}
\title{World population data}
\format{
A data frame with 235 observations with 72 columns.
\describe{
\item{CountryName}{The name of the observed country;  Unit: none;         Datatype: chr}
\item{1950:2020}{Each individual year's population;   Unit: 1,000 people; Datatype: num}
}
}
\usage{
World_Population
}
\description{
A set of data about the the world population from 1950 to 2020.
}
\keyword{datasets}
